+++ ouroboros-network/sim-tests/Test/Ouroboros/Network/Testnet.hs
@@ -16,7 +16,6 @@ import           Control.Monad.Class.MonadFork
 import           Control.Monad.Class.MonadTime.SI (DiffTime, Time (Time),
                      addTime, diffTime)
 import           Control.Monad.IOSim
-import           Control.Tracer (Tracer (Tracer))
 import           Data.Bifoldable (bifoldMap)
 
 import           Data.List (find, intercalate, tails)
@@ -88,7 +87,7 @@ import           Ouroboros.Network.PeerSelection.PeerAdvertise
                      (PeerAdvertise (..))
 import           Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import           Ouroboros.Network.PeerSelection.State.LocalRootPeers
-                     (HotValency (..))
+                     (HotValency (..), WarmValency (..))
 
 tests :: TestTree
 tests =
@@ -174,14 +173,6 @@ tests =
 #endif
   ]
 
-
-tracerDiffusionSimWithTimeName :: Tracer (IOSim s) (WithName NtNAddr DiffusionSimulationTrace)
-tracerDiffusionSimWithTimeName = tracerWithTime tracer
-  where
-    tracer :: Tracer (IOSim s) (WithTime (WithName NtNAddr DiffusionSimulationTrace))
-    tracer = dynamicTracer
-
-
 -- | This test coverage of ConnectionManagerTrace constructors.
 --
 prop_connection_manager_trace_coverage :: AbsBearerInfo
@@ -193,7 +184,6 @@ prop_connection_manager_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [ConnectionManagerTrace
                   NtNAddr
@@ -229,7 +219,6 @@ prop_connection_manager_transitions_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [AbstractTransitionTrace NtNAddr]
       events = mapMaybe (\case DiffusionConnectionManagerTransitionTrace st ->
@@ -266,7 +255,6 @@ prop_inbound_governor_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [InboundGovernorTrace NtNAddr]
       events = mapMaybe (\case DiffusionInboundGovernorTrace st -> Just st
@@ -299,7 +287,6 @@ prop_inbound_governor_transitions_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [RemoteTransitionTrace NtNAddr]
       events = mapMaybe (\case DiffusionInboundGovernorTransitionTrace st ->
@@ -335,7 +322,6 @@ prop_fetch_client_state_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [TraceFetchClientState BlockHeader]
       events = mapMaybe (\case DiffusionFetchTrace st ->
@@ -443,7 +429,6 @@ prop_server_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [ServerTrace NtNAddr]
       events = mapMaybe (\case DiffusionServerTrace st -> Just st
@@ -476,7 +461,6 @@ prop_peer_selection_action_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [PeerSelectionActionsTrace NtNAddr NtNVersion]
       events = mapMaybe (\case DiffusionPeerSelectionActionsTrace st -> Just st
@@ -520,7 +504,6 @@ prop_peer_selection_trace_coverage defaultBearerInfo diffScript =
       sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                 diffScript
                                 iosimTracer
-                                tracerDiffusionSimWithTimeName
 
       events :: [TracePeerSelection NtNAddr]
       events = mapMaybe (\case DiffusionPeerSelectionTrace st -> Just st
@@ -659,7 +642,6 @@ prop_diffusion_nolivelock defaultBearerInfo diffScript@(DiffusionScript _ _ l) =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         trace :: [(Time, ThreadId (IOSim s), Maybe ThreadLabel, SimEventType)]
         trace = take 125000
@@ -729,7 +711,6 @@ prop_diffusion_dns_can_recover defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -791,8 +772,7 @@ prop_diffusion_dns_can_recover defaultBearerInfo diffScript =
 
     verify_dns_can_recover :: Events DiffusionTestTrace -> Property
     verify_dns_can_recover events =
-        counterexample (intercalate "\n" $ map show $ eventsToList events)
-      $ verify Map.empty Map.empty 0 (Time 0) (Signal.eventsToList events)
+        verify Map.empty Map.empty 0 (Time 0) (Signal.eventsToList events)
 
     verify :: Map DNS.Domain Time
            -> Map DNS.Domain DiffTime
@@ -937,7 +917,6 @@ prop_diffusion_target_established_public defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1031,7 +1010,6 @@ prop_diffusion_target_active_public defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1113,7 +1091,6 @@ prop_diffusion_target_active_local defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1196,7 +1173,6 @@ prop_diffusion_target_active_root defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1318,7 +1294,6 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1354,12 +1329,9 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
   where
     verify_target_established_local :: Events DiffusionTestTrace -> Property
     verify_target_established_local events =
-      let govLocalRootPeersSig :: Signal (Set NtNAddr)
+      let govLocalRootPeersSig :: Signal (LocalRootPeers.LocalRootPeers NtNAddr)
           govLocalRootPeersSig =
-            selectDiffusionPeerSelectionState
-              ( LocalRootPeers.keysSet
-              . Governor.localRootPeers)
-              events
+            selectDiffusionPeerSelectionState Governor.localRootPeers events
 
           govInProgressPromoteColdSig :: Signal (Set NtNAddr)
           govInProgressPromoteColdSig =
@@ -1435,10 +1407,17 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
           promotionOpportunities :: Signal (Set NtNAddr)
           promotionOpportunities =
             (\local established recentFailures inProgressPromoteCold isAlive ->
-              if isAlive
-              then local Set.\\ established
-                         Set.\\ recentFailures
-                         Set.\\ inProgressPromoteCold
+              if isAlive then
+                Set.unions
+                  [ -- There are no opportunities if we're at or above target
+                    if Set.size groupEstablished >= warmTarget
+                       then Set.empty
+                       else groupEstablished Set.\\ established
+                                             Set.\\ recentFailures
+                                             Set.\\ inProgressPromoteCold
+                  | (_, WarmValency warmTarget, group) <- LocalRootPeers.toGroupSets local
+                  , let groupEstablished = group `Set.intersection` established
+                  ]
               else Set.empty
             ) <$> govLocalRootPeersSig
               <*> govEstablishedPeersSig
@@ -1448,7 +1427,7 @@ prop_diffusion_target_established_local defaultBearerInfo diffScript =
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
-            Signal.keyedTimeout
+            Signal.keyedTimeoutTruncated
               15 -- seconds
               id
               promotionOpportunities
@@ -1483,7 +1462,6 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1493,6 +1471,8 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
                . splitWithNameTrace
                . Trace.fromList ()
                . fmap snd
+               . Signal.eventsToList
+               . Signal.eventsFromListUpToTime (Time (10 * 60 * 60))
                . Trace.toList
                . fmap (\(WithTime t (WithName name b)) -> (t, WithName name (WithTime t b)))
                . withTimeNameTraceEvents
@@ -1523,15 +1503,24 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
           govLocalRootPeersSig =
             selectDiffusionPeerSelectionState Governor.localRootPeers events
 
+          govActiveTargetsSig :: Signal Int
+          govActiveTargetsSig =
+            selectDiffusionPeerSelectionState
+              (targetNumberOfActivePeers . Governor.targets)
+              events
+
           govEstablishedPeersSig :: Signal (Set NtNAddr)
           govEstablishedPeersSig =
             selectDiffusionPeerSelectionState
-              (EstablishedPeers.toSet . Governor.establishedPeers)
+              (dropBigLedgerPeers $
+                 EstablishedPeers.toSet . Governor.establishedPeers)
               events
 
           govActivePeersSig :: Signal (Set NtNAddr)
           govActivePeersSig =
-            selectDiffusionPeerSelectionState Governor.activePeers events
+            selectDiffusionPeerSelectionState
+              (dropBigLedgerPeers Governor.activePeers)
+              events
 
           govActiveFailuresSig :: Signal (Set NtNAddr)
           govActiveFailuresSig =
@@ -1556,10 +1545,21 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
                          | otherwise         -> Just failures
                          where
                            failures = Map.keysSet (Map.filter (==PeerWarm) . fmap fst $ status)
+                       TracePromoteWarmBigLedgerPeerFailed _ _ peer _ ->
+                         Just (Set.singleton peer)
+                       TraceDemoteBigLedgerPeersAsynchronous status
+                         | Set.null failures -> Nothing
+                         | otherwise -> Just failures
+                         where
+                           failures = Map.keysSet (Map.filter ((==PeerCold) . fst) status)
                        _ -> Nothing
                 )
             . selectDiffusionPeerSelectionEvents
             $ events
+
+          govInProgressPromoteWarmSig :: Signal (Set NtNAddr)
+          govInProgressPromoteWarmSig =
+            selectDiffusionPeerSelectionState Governor.inProgressPromoteWarm events
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
@@ -1588,22 +1588,31 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
                                   (const False)
                                   trJoinKillSig
 
+          -- There are no opportunities if we're at or above target.
+          --
+          -- We define local root peers not to be promotion opportunities for
+          -- the purpose of the general target of active peers.
+          -- The local root peers have a separate target with a separate property.
+          -- And we cannot count local peers since we can have corner cases where
+          -- the only choices are local roots in a group that is already at target
+          -- but the general target is still higher. In such situations we do not
+          -- want to promote any, since we'd then be above target for the local
+          -- root peer group.
+          --
+          promotionOpportunity target local established active recentFailures isAlive
+            | isAlive && Set.size active < target
+            = established Set.\\ active
+                          Set.\\ LocalRootPeers.keysSet local
+                          Set.\\ recentFailures
+
+            | otherwise
+            = Set.empty
+
           promotionOpportunities :: Signal (Set NtNAddr)
           promotionOpportunities =
-            (\local established active recentFailures isAlive ->
-              if isAlive
-              then Set.unions
-                    [ -- There are no opportunities if we're at or above target
-                      if Set.size groupActive >= hotTarget
-                         then Set.empty
-                         else groupEstablished Set.\\ active
-                                               Set.\\ recentFailures
-                    | (HotValency hotTarget, _, group) <- LocalRootPeers.toGroupSets local
-                    , let groupActive      = group `Set.intersection` active
-                          groupEstablished = group `Set.intersection` established
-                    ]
-              else Set.empty
-            ) <$> govLocalRootPeersSig
+            promotionOpportunity
+              <$> govActiveTargetsSig
+              <*> govLocalRootPeersSig
               <*> govEstablishedPeersSig
               <*> govActivePeersSig
               <*> govActiveFailuresSig
@@ -1611,7 +1620,7 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
-            Signal.keyedTimeout
+            Signal.keyedTimeoutTruncated
               10 -- seconds
               id
               promotionOpportunities
@@ -1619,10 +1628,19 @@ prop_diffusion_target_active_below defaultBearerInfo diffScript =
        in counterexample
             ("\nSignal key: (local, established peers, active peers, " ++
              "recent failures, opportunities, ignored too long)") $
+          counterexample
+            (intercalate "\n" $ map show $ Signal.eventsToList events) $
 
           signalProperty 20 show
-            (\toolong -> Set.null toolong)
-            promotionOpportunitiesIgnoredTooLong
+            (\(_, _, _, _, _, _, toolong) -> Set.null toolong)
+            ((,,,,,,) <$> govLocalRootPeersSig
+                 <*> govEstablishedPeersSig
+                 <*> govActivePeersSig
+                 <*> govActiveFailuresSig
+                 <*> govInProgressPromoteWarmSig
+                 <*> trIsNodeAlive
+                 <*> promotionOpportunitiesIgnoredTooLong
+            )
 
 
 prop_diffusion_target_active_local_below :: AbsBearerInfo
@@ -1633,7 +1651,6 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1652,7 +1669,7 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
                   @NtNAddr
                . Trace.fromList (MainReturn (Time 0) () [])
                . fmap (\(t, tid, tl, te) -> SimEvent t tid tl te)
-               . take 125000
+               . take 250000
                . traceEvents
                $ runSimTrace sim
 
@@ -1765,7 +1782,7 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
 
           promotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           promotionOpportunitiesIgnoredTooLong =
-            Signal.keyedTimeout
+            Signal.keyedTimeoutTruncated
               10 -- seconds
               id
               promotionOpportunities
@@ -1773,6 +1790,8 @@ prop_diffusion_target_active_local_below defaultBearerInfo diffScript =
        in counterexample
             ("\nSignal key: (local, established peers, active peers, " ++
              "recent failures, opportunities, ignored too long)") $
+          counterexample
+            (intercalate "\n" $ map show $ Signal.eventsToList events) $
 
           signalProperty 20 show
             (\(_,_,_,_,_,toolong) -> Set.null toolong)
@@ -1865,7 +1884,6 @@ prop_diffusion_async_demotions defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -1949,7 +1967,7 @@ prop_diffusion_async_demotions defaultBearerInfo diffScript =
 
           demotionOpportunitiesTooLong :: Signal (Set NtNAddr)
           demotionOpportunitiesTooLong =
-              Signal.keyedTimeout 1 id demotionOpportunities
+              Signal.keyedTimeoutTruncated 1 id demotionOpportunities
 
       in signalProperty
             20 show Set.null
@@ -1978,7 +1996,6 @@ prop_diffusion_target_active_local_above defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Events DiffusionTestTrace]
         events = fmap ( Signal.eventsFromList
@@ -2068,8 +2085,8 @@ prop_diffusion_target_active_local_above defaultBearerInfo diffScript =
 
           demotionOpportunitiesIgnoredTooLong :: Signal (Set NtNAddr)
           demotionOpportunitiesIgnoredTooLong =
-            Signal.keyedTimeout
-              15 -- seconds
+            Signal.keyedTimeoutTruncated
+              50 -- seconds
               id
               demotionOpportunities
 
@@ -2100,7 +2117,6 @@ prop_diffusion_cm_valid_transitions defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2201,7 +2217,6 @@ prop_diffusion_cm_valid_transition_order defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2344,7 +2359,6 @@ prop_diffusion_cm_no_dodgy_traces defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2407,7 +2421,6 @@ prop_diffusion_peer_selection_actions_no_dodgy_traces defaultBearerInfo (HotDiff
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   (DiffusionScript sa dns hds)
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2660,7 +2673,6 @@ prop_diffusion_ig_valid_transitions defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2730,7 +2742,6 @@ prop_diffusion_ig_valid_transition_order defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (WithName NtNAddr (WithTime DiffusionTestTrace))]
         events = fmap (Trace.fromList ())
@@ -2756,7 +2767,8 @@ prop_diffusion_ig_valid_transition_order defaultBearerInfo diffScript =
             lastTime = (\(WithName _ (WithTime t _)) -> t)
                      . last
                      $ evsList
-         in classifySimulatedTime lastTime
+         in counterexample (Trace.ppTrace show show ev)
+          $ classifySimulatedTime lastTime
           $ classifyNumberOfEvents (length evsList)
           $ verify_ig_valid_transition_order
           $ (\(WithName _ (WithTime _ b)) -> b)
@@ -2797,7 +2809,6 @@ prop_diffusion_timeouts_enforced defaultBearerInfo diffScript =
         sim = diffusionSimulation (toBearerInfo defaultBearerInfo)
                                   diffScript
                                   iosimTracer
-                                  tracerDiffusionSimWithTimeName
 
         events :: [Trace () (Time, DiffusionTestTrace)]
         events = fmap ( Trace.fromList ()
@@ -2876,9 +2887,6 @@ classifyNumberOfEvents nEvents =
       . classify (nEvents >= 1000) "Nº Events >= 1000"
       . classify (nEvents >= 10000) "Nº Events >= 10000"
       . classify (nEvents >= 50000) "Nº Events >= 50000"
-
-dynamicTracer :: (Typeable a, Show a) => Tracer (IOSim s) a
-dynamicTracer = Tracer traceM <> sayTracer
 
 withTimeNameTraceEvents :: forall b name r. (Typeable b, Typeable name)
                         => SimTrace r
@@ -3017,3 +3025,11 @@ labelDiffusionScript (DiffusionScript args _ nodes) =
 
     -- TODO: it would be nice to check if the graph is connected if all dns
     -- names can be resolved.
+
+-- | filter out big ledger peers
+--
+dropBigLedgerPeers
+    :: (Governor.PeerSelectionState NtNAddr peerconn -> Set NtNAddr)
+    ->  Governor.PeerSelectionState NtNAddr peerconn -> Set NtNAddr
+dropBigLedgerPeers f =
+  \st -> f st Set.\\ Governor.bigLedgerPeers st